:app — Settings root subtitles + fold API keys into Voice
Each Settings root row now carries a one-line subtitle previewing what
lives inside ("Notifications and Times", "Temperatures and Outfits",
"Languages and Units", "Providers and Accents", "Location and Calendar").
SettingsNavRow grows an optional `subtitle` and SettingsRoute carries a
`subtitleRes` so the mapping lives next to each route.

API keys move out of the root list and into Voice. The prose itself is
generated locally now (InsightFormatter), so all three BYOK keys —
Gemini, OpenAI, ElevenLabs — only gate cloud TTS. Voice already showed
inline key entry as a convenience; promote that to full management
(status + save + clear) for the currently selected engine via the
existing `KeyEntryFields` helper, drop the now-redundant `ApiKeys`
sub-page and `MissingKeyHint` / `AddApiKeyDialog` helpers, and rename
the now-single-symbol file to `KeyEntryFields.kt`.

Privacy: nothing new leaves the device.
