+++ app/src/main/res/values-de/strings.xml
@@ -37,7 +37,6 @@ zh-CN) is unchanged; only the displayed label localizes.
     <string name="settings_root_schedule">Zeitplan</string>
     <string name="settings_root_region">Region</string>
     <string name="settings_root_clothes">Kleidung</string>
-    <string name="settings_root_api_keys">API-Schlüssel</string>
     <string name="settings_root_data_sources">Datenquellen</string>
     <string name="settings_root_about">Über</string>
