+++ app/src/main/res/values-ja/strings.xml
@@ -48,7 +48,6 @@ the displayed label localizes.
     <string name="settings_root_schedule">スケジュール</string>
     <string name="settings_root_region">地域</string>
     <string name="settings_root_clothes">服装</string>
-    <string name="settings_root_api_keys">API キー</string>
     <string name="settings_root_data_sources">データソース</string>
     <string name="settings_root_about">アプリについて</string>
