+++ app/src/main/res/values/strings.xml
@@ -72,9 +72,14 @@
     <string name="settings_root_schedule">Schedule</string>
     <string name="settings_root_region">Region</string>
     <string name="settings_root_clothes">Clothes</string>
-    <string name="settings_root_api_keys">API keys</string>
     <string name="settings_root_data_sources">Data sources</string>
     <string name="settings_root_about">About</string>
+
+    <string name="settings_root_schedule_subtitle">Notifications and Times</string>
+    <string name="settings_root_clothes_subtitle">Temperatures and Outfits</string>
+    <string name="settings_root_region_subtitle">Languages and Units</string>
+    <string name="settings_root_voice_subtitle">Providers and Accents</string>
+    <string name="settings_root_data_sources_subtitle">Location and Calendar</string>
 
     <string name="settings_api_keys_title">API keys</string>
     <string name="settings_api_keys_description">Configure the API keys for the online voice engines you want to use. Keys are encrypted at rest using the Android Keystore.</string>
